combine iosTokenCacheAccessor and iOSlegacy stuff
