+++ src/Microsoft.IdentityModel.Clients.ActiveDirectory/AuthenticationContext.cs
@@ -182,7 +182,6 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory
         }
 
 #if iOS
-
         private string keychainSecurityGroup;
 
         /// <summary>
@@ -201,7 +200,6 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory
             set
             {
                 keychainSecurityGroup = value;
-                StorageDelegates.LegacyCachePersistence.SetKeychainSecurityGroup(value);
                 TokenCache.TokenCacheAccessor.SetiOSKeychainSecurityGroup(value);
             }
         }
@@ -223,7 +221,6 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory
             set
             {
                 keychainSecurityGroup = value;
-                StorageDelegates.LegacyCachePersistence.SetKeychainSecurityGroup(value);
                 TokenCache.TokenCacheAccessor.SetKeychainSecurityGroup(value);
             }
         }
