+++ src/Microsoft.IdentityModel.Clients.ActiveDirectory/Platforms/iOS/iOSPlatformProxy.cs
@@ -31,7 +31,6 @@ using System.Threading.Tasks;
 using Microsoft.Identity.Core.Cache;
 using UIKit;
 using Foundation;
-using Microsoft.Identity.Core.Http;
 
 namespace Microsoft.Identity.Core
 {
@@ -127,12 +126,12 @@ namespace Microsoft.Identity.Core
 
         public ILegacyCachePersistence CreateLegacyCachePersistence()
         {
-            return new iOSLegacyCachePersistence();
+            return new iOSTokenCacheAccessor();
         }
 
         public ITokenCacheAccessor CreateTokenCacheAccessor()
         {
-            return new iOSTokenCacheAccessor();
+            return new iOSTokenCacheAccessor() as ITokenCacheAccessor;
         }
 
         /// <inheritdoc />
